Add udp client and server.

+++ Common/Network/Address.cs
@@ -3,5 +3,5 @@
 public record Address
 {
     public string Ip { get; set; } = "127.0.0.1";
-    public ushort Port { get; set; } = 0;
+    public int Port { get; set; } = 0;
 }
